refactor: move drag lifetime guards into abstract
