+++ investigations/collision-observer/IMPLEMENTATION.md
@@ -31,7 +31,11 @@ Collision selection and action completion have separate private state. The obser
 
 `move()` retains its existing queued default position write. Returning an asynchronous handler does not delay that write or replay old movement when the promise settles. A handler that consumes movement must prevent the default before the queued write. The action remains pending until its default write, returned work, and rendering finish. Normal drop drains accepted actions and reconciles final collisions before taking its snapshot; new movement is refused after stop is requested. Cancellation remains immediate, and stale completions cannot release a later operation's work.
 
-Optimistic sorting returns its placement work and source acknowledgment from an ordinary `Plugin` event handler. The DOM renderer boundary measures the committed layout before action completion; sortable plugins do not collect affected rows or old/new ancestors. The abstract action layer sequences relative input before dispatch, owns queued commands immediately, and discards undispatched commands on cancellation. Absolute pointer input bypasses that queue. Keyboard handlers return one promise per command and await only their own target action; they maintain no private input queue. They update position as part of that original input, so no privileged continuation or second synthetic `dragmove` is needed. Pending tasks are guarded against cancellation, replacement, and destruction inside the DOM package. A private DOM core plugin adapts the existing renderer so its completion includes geometry measurement. The same behavior covers ordinary asynchronous monitor listeners, including move handlers, without a sortable-specific task wrapper.
+Optimistic sorting returns its placement work and source acknowledgment from an ordinary `Plugin` event handler. The DOM renderer boundary measures the committed layout before action completion; sortable plugins do not collect affected rows or old/new ancestors. The abstract action layer sequences relative input before dispatch, owns queued commands immediately, and discards undispatched commands on cancellation. Absolute pointer input bypasses that queue. Keyboard handlers return one promise per command and await only their own target action; they maintain no private input queue. They update position as part of that original input, so no privileged continuation or second synthetic `dragmove` is needed.
+
+Pending tasks are guarded against cancellation, replacement, and destruction by a private abstract utility. That utility depends only on abstract drag state and `AbortController`; its lifetime and plugin-completion tests live in abstract. DOM sorting bundles include the helper through a build-time source import. It has no shared module state, no runtime source-file dependency, and no public export or new `Plugin` member.
+
+A private DOM core plugin adapts the existing renderer so its completion includes geometry measurement. The same behavior covers ordinary asynchronous monitor listeners, including move handlers, without a sortable-specific task wrapper.
 
 There is no collision-specific plugin subclass, observer capability, shared collision-event gate, public transaction API, or cross-package access to private state. Abstract code knows only actions, events, returned promises, and rendering. DOM ownership and measurement stay in the DOM package.
